% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEA_utils.R
\name{calc_LFC_scmat}
\alias{calc_LFC_scmat}
\title{Calculate Log Fold Change (LFC) from Single-Cell Metabolomics Matrix}
\usage{
calc_LFC_scmat(object)
}
\arguments{
\item{object}{A \code{S2IsoMEr} object that contains the single-cell metabolomics matrix (\code{scmatrix}) and metadata about conditions (\code{conditions}, \code{condition.x}, and \code{condition.y}).}
}
\value{
A numeric vector of Log Fold Change (LFC) values for each metabolite. The length of the vector matches the number of metabolites in the matrix.
}
\description{
This method calculates the Log Fold Change (LFC) between two conditions in a single-cell metabolomics matrix. It compares the mean metabolite expression levels between the specified conditions and returns the LFC values for each metabolite.
}
\details{
This method computes the mean expression levels of metabolites for two conditions specified in the \code{S2IsoMEr} object. The LFC is calculated as the difference in log2-transformed mean expression levels between the two conditions.
}
\examples{
\dontrun{
data("example_ORA_obj")
lfc_values <- calc_LFC_scmat(example_ORA_obj)
}

}
