% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORA_decoupler.R
\name{decouple_ORA_wrapper}
\alias{decouple_ORA_wrapper}
\title{Run ORA Analysis with Optional Bootstrapping}
\usage{
decouple_ORA_wrapper(
  marker_list,
  term_list,
  universe,
  pass_adjust = F,
  seed = 42,
  ORA_boot = T
)
}
\arguments{
\item{marker_list}{A list of metabolites per condition. If there are no conditions, please provide it as list("condition" = c("metabo_1", "metabo_2", ...))}

\item{term_list}{A list of named terms where each term is a character vector of metabolites.}

\item{universe}{A character vector specifying the universe of metabolites.}

\item{pass_adjust}{A logical indicating whether to pass the adjustment parameter to the ORA function. If TRUE, the contingency table won't be adjusted. Only required for bootstrap-ORA.}

\item{seed}{An integer specifying the seed for reproducibility.}

\item{ORA_boot}{A logical indicating whether to use bootstrapped ORA. If TRUE, the function will use the bootstrapped ORA method; otherwise, it will use a simple ORA method.}
}
\value{
A list containing two elements:
\item{ORA_res}{A data frame with the ORA results, including columns for statistic, source, condition, score, and p-value.}
\item{ORA_conting}{A data frame with the ORA contingency table, including columns for source, condition, TP, FP, FN, and TN.}
}
\description{
This function performs over-representation analysis (ORA) on a given list of markers using either a bootstrapped or a simple method.
It is a wrapper around the \code{run_ora} function from the \code{decoupleR} package. This function is also used internally in the \code{Run_bootstrap_ORA} and \code{Run_simple_ORA} functions.
}
\details{
This function is a wrapper around the \code{run_ora} function from the \code{decoupleR} package. For more details, please refer to the \href{https://saezlab.github.io/decoupleR/reference/run_ora.html}{decoupleR documentation}.
}
\examples{
\dontrun{
marker_list <- c("B2", "B5", "B8")
term_list <- list(
  "Term1" = c("B1", "B2", "B3", "B4"),
  "Term2" = c("B4", "B5", "B6", "B9")
)
universe <- c("B1", "B2", "B3", "B4", "B5", "B6", "B7", "B8", "B9")
results <- decouple_ORA_wrapper(marker_list, term_list, universe, pass_adjust = TRUE, seed = 42, ORA_boot = FALSE)
}
}
\references{
Badia-i-Mompel, P., Nagai, J. S., & Saez-Rodriguez, J. (2022). decoupleR: A flexible tool to handle various modes of biological network analysis. \emph{Bioinformatics Advances}, 2(1), vbac016. \href{https://academic.oup.com/bioinformaticsadvances/article/2/1/vbac016/6544613?login=true}{https://doi.org/10.1093/bioadv/vbac016}
}
