% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_MSEA_multicond}
\alias{example_MSEA_multicond}
\title{Example Multi condition MSEA results}
\format{
A data frame with 100 rows and 9 variables:
\describe{
\item{Term}{Character: Term name of metabolite set (e.g. Triacylglycerols)}
\item{n}{Numeric: Median term/query overlap size over bootstraps.}
\item{ES_median}{Numeric: Median of NES (normalized enrichment score) for fgsea or ES for ks_signed}
\item{ES_sd}{Numeric: stadndard deviation of NES (normalized enrichment score) for fgsea or ES for ks_signed}
\item{p.value_combined}{Numeric: Combined p-value using \code{metap::sumlog}}
\item{q.value_combined}{Numeric: Combined adjusted p-value using \code{metap::sumlog}}
\item{fraction.bootstrap.presence}{Numeric: Proportion of bootstraps the given term was tested}
\item{condition.x}{Character: Reference condition}
\item{condition.y}{Character: Query condition}
}
}
\usage{
data(example_MSEA_multicond)
}
\description{
The \code{example_MSEA_multicond} dataset is a data frame that contains results of running \code{\link{Run_bootstrap_MSEA}} on multiple pairwise conditions compared to same reference.
}
\examples{
# Load the example data frame
data(example_MSEA_multicond)

# View the first few rows
head(example_MSEA_multicond)


}
\keyword{datasets}
