% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_ORA_custom_universe}
\alias{example_ORA_custom_universe}
\title{Example custom universe for ORA}
\format{
A character vector with 1618 elements, each representing an ion (sum formula + adduct)
}
\usage{
data(example_ORA_custom_universe)
}
\description{
\code{example_ORA_custom_universe} is a character vector of ions representing a custom universe for ORA
}
\examples{
# Load the example vector
data(example_ORA_custom_universe)

# View the first few elements
head(example_ORA_custom_universe)
}
\keyword{datasets}
