% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_ORA_markers}
\alias{example_ORA_markers}
\title{Example List of Markers for ORA}
\format{
A list with the following elements:
\describe{
\item{upregulated}{A character vector of upregulated ions}
\item{downregulated}{A character vector of downregulated ions}
\item{all}{A character vector of all markers, including both upregulated and downregulated}
}
}
\usage{
data(example_ORA_markers)
}
\description{
\code{example_ORA_markers} is a list that provides markers for use in ORA
}
\examples{
# Load the marker list
data(example_ORA_markers)

}
\keyword{datasets}
