% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORA_utils.R
\name{get_TP_markers_per_Term}
\alias{get_TP_markers_per_Term}
\title{Get intersection between term and query for a Specific Term}
\usage{
get_TP_markers_per_Term(ORA_boot_df, term_of_interest)
}
\arguments{
\item{ORA_boot_df}{A data frame containing the ORA bootstrap results.}

\item{term_of_interest}{A string specifying the term for which true positive markers should be extracted.}
}
\value{
A character vector containing the unique true positive markers associated with the specified term.
}
\description{
This function extracts unique true positive markers for a specified term from the ORA bootstrap result.
}
\examples{
\dontrun{
  # Example usage with hypothetical data
  ORA_boot_df <- data.frame(
    Term = c("Term1", "Term2", "Term1"),
    TP_markers = c("Marker1;Marker2", "Marker3;Marker4", "Marker2;Marker5")
  )
  markers <- get_TP_markers_per_Term(ORA_boot_df, "Term1")
  print(markers)
}

}
