% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORA_utils.R
\name{map_TP_markers_to_ions}
\alias{map_TP_markers_to_ions}
\title{Map True Positive Markers to Input Ions from single cell matrix.}
\usage{
map_TP_markers_to_ions(markers, scm_ions)
}
\arguments{
\item{markers}{A character vector containing true positive markers obtained from \code{\link{get_TP_markers_per_Term}}}

\item{scm_ions}{A character vector containing input ion names.}
}
\value{
A character vector containing the ions that match the given true positive markers.
}
\description{
This function maps true positive markers to their corresponding ions by modifying and matching the marker and ion names.
}
\examples{
markers <- c("sf.Na", "sf.H")
scm_ions <- c("sf+Na", "sf+H")
map_TP_markers_to_ions(markers, scm_ions)

}
