% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORA_utils.R
\name{metabo_bootstrap}
\alias{metabo_bootstrap}
\title{Perform Bootstrap Sampling on Metabolite Annotations}
\usage{
metabo_bootstrap(annot_list, annot_weights = NULL, n_bootstraps = 50)
}
\arguments{
\item{annot_list}{A list of character vectors where each vector contains metabolites or molecules for a particular annotation. Each element in the list corresponds to a specific sum formula.}

\item{annot_weights}{A list of numeric vectors where each vector contains weights associated with the metabolites in \code{annot_list}. Must have the same length as \code{annot_list}. If NULL, all metabolites are equally weighted.}

\item{n_bootstraps}{An integer specifying the number of bootstrap samples to generate.}
}
\value{
A list of length \code{n_bootstraps}, where each element is a list of bootstrap samples for each sum formula. Each bootstrap sample is a character vector of sampled metabolites.
}
\description{
This function performs bootstrap sampling on metabolite annotations, allowing for resampling of molecules with or without associated weights. It generates multiple bootstrap samples for further analysis.
}
\details{
This function is used to create multiple bootstrap samples of metabolite annotations by randomly sampling metabolites from each sum formula. If weights are provided, the sampling is performed according to these weights; otherwise, the sampling is uniform.
}
\examples{
\dontrun{
annot_list <- list(
  Sf1 = c("Metabolite1", "Metabolite2", "Metabolite3"),
  Sf2 = c("MetaboliteA", "MetaboliteB")
)
annot_weights <- list(
  Sf1 = c(0.1, 0.2, 0.7),
  Sf2 = c(0.5, 0.5)
)
bootstrapped_samples <- metabo_bootstrap(annot_list, annot_weights, n_bootstraps = 100)
}
}
