% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R, R/enrichment_functions.r
\name{plotEnrichment}
\alias{plotEnrichment}
\title{Plot bootstrap enrichment analysis}
\usage{
plotEnrichment(object, ...)

plotEnrichment(object, ...)
}
\arguments{
\item{object}{A bmetenrichr object after enrichment analysis.}

\item{min.annotations}{An integer describing the minimal number of annotations each term should include}

\item{q.value.cutoff}{A numeric between 0 and 1. Only terms with q-values lower than this value will be displayed.}

\item{bootstrap.fraction.cutoff}{A numeric between 0 and 1 (default = 0.5), indicating the minimal fraction that the metabolite set is present in all bootstrap iterations.}

\item{plotIDs}{A logical indicating whether term IDs should be displayed.}

\item{by.statistic}{A character indicating how the x-axis will be arranged.
Can be either 'ES' (enrichment score) or 'q.value' (default).}
}
\value{
A ggplot2 object.

A ggplot2 object.
}
\description{
Plot bootstrap enrichment analysis

Plot bootstrap enrichment analysis
}
\examples{

plotEnrichment(myTestRun)


plotEnrichment(myTestRun)https://www.airbnb.com/rooms/936205035669822515?c=.pi80.pkYm9va2luZy92Ml9taWdyYXRpb24vcmVzZXJ2YXRpb25fZ3Vlc3RfY29uZmlybWF0aW9u&euid=7ac519aa-e54c-ddf7-9992-73cad5af86cd

}
