% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_funcs.R
\name{simple_fgsea}
\alias{simple_fgsea}
\title{Run simple fGSEA enrichment}
\usage{
simple_fgsea(
  pathways,
  stats,
  minSize = 1,
  maxSize = length(stats) - 1,
  eps = 1e-50,
  scoreType = c("std", "pos", "neg"),
  nPermSimple = 1000
)
}
\arguments{
\item{pathways}{List of metabolite sets to check.}

\item{stats}{Named vector of metabolite-level stats. Names should be the same as in 'pathways'}

\item{minSize}{Minimal size of a metabolite set to test. All pathways below the threshold are excluded.}

\item{maxSize}{Maximal size of a metabolite set to test. All pathways above the threshold are excluded.}

\item{eps}{This parameter sets the boundary for calculating the p value.}

\item{scoreType}{This parameter defines the GSEA score type. Possible options are ("std", "pos", "neg")}

\item{nPermSimple}{Number of permutations in the simple fgsea implementation for preliminary estimation of P-values}
}
\value{

}
\description{
simple_fgsea calls a simple fGSEA from the fgsea package
}
