% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{barplot_ORA_boot}
\alias{barplot_ORA_boot}
\title{Bar Plot for ORA Bootstrapped Results}
\usage{
barplot_ORA_boot(ORA_boot_res)
}
\arguments{
\item{ORA_boot_res}{A list containing the ORA bootstrapped results. The list should include \code{unfiltered_enrich_res} and \code{clean_enrich_res} data frames.}
}
\value{
A ggplot object displaying the bar plot of ORA bootstrapped results.
}
\description{
This function creates a bar plot for Over-Representation Analysis (ORA) bootstrapped results, including error bars to show the range of enrichment scores (OR) and indicating the significance based on q-values.
}
\details{
The function creates a bar plot where the terms are reordered by their median enrichment scores. Bars are filled with colors representing the combined q-values, and error bars show the minimum and maximum enrichment scores (ES_min and ES_max) across bootstraps. The number of true positives (n) is included in the term labels.
}
\examples{
\dontrun{
  plot <- barplot_ORA_boot(ORA_boot_res = ora_boot_results)
  print(plot)
}

}
