% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{compare_metabo_distr}
\alias{compare_metabo_distr}
\title{Plot intensity distribution of metabolite across conditions}
\usage{
compare_metabo_distr(obj, metabolite, conds_of_interest = NULL)
}
\arguments{
\item{metabolite}{A character indicating metabolite name}

\item{object}{A bmetenrichr object after enrichment analysis.}
}
\value{
A ggplot2 object.
}
\description{
Plot intensity distribution of metabolite across conditions
}
\examples{

compare_metabo_distr(obj, metabolite="C6H8N2O3-H")

}
