% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{compare_metabo_distr}
\alias{compare_metabo_distr}
\title{Plot intensity distribution of metabolite across conditions}
\usage{
compare_metabo_distr(obj, metabolite, conds_of_interest = NULL)
}
\arguments{
\item{obj}{A S2IsoMEr object after enrichment analysis.}

\item{metabolite}{A character indicating metabolite name}

\item{conds_of_interest}{Optional. A character vector specifying which conditions to include in the plot. If \code{NULL} (the default), all conditions are included.}
}
\value{
A \code{ggplot2} object displaying the density plot of metabolite intensities across conditions.
}
\description{
Plot intensity distribution of metabolite across conditions
}
\examples{

compare_metabo_distr(obj, metabolite="C6H8N2O3-H")

}
