% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{initEnrichment}
\alias{initEnrichment}
\title{Generate S2IsoMEr enrichment object}
\usage{
initEnrichment(
  scmatrix,
  conditions,
  enrichment_type = c("ORA", "MSEA"),
  annot_db = "HMDB",
  annot_custom_db = NULL,
  endogenous_only = T,
  pathway_assoc_only = F,
  remove_expected_predicted = T,
  annotations = NULL,
  annotation.weights = NULL,
  consider_isomers = TRUE,
  consider_isobars = FALSE,
  mass_range_ppm = 3,
  polarization_mode = "positive",
  include = NULL,
  molecule_type = c("Lipid", "Metabo"),
  background_type = c("LION", "main_class", "super_class", "sub_class", "pathways"),
  custom_bg = NULL,
  termsOfInterest = "all",
  condition.x = NULL,
  condition.y = NULL,
  ranking.by = c("wilcox.test", "t.test", "BWS", "logFC"),
  gsea.method = c("fgsea", "ks_signed")
)
}
\arguments{
\item{scmatrix}{A numeric matrix of n metabolites (rows) and m cells or measurments (columns).}

\item{conditions}{A vector of length m with condition identifiers.}

\item{enrichment_type}{A character specifying whether Overrepresentation analysis (ORA) or Metabolite Set Enrichment Analysis (MSEA) is performed}

\item{annot_db}{A character or character vector specifying which annotation database(s) were used. Current databases include ("CoreMetabolome", "HMDB","SwissLipids","LipidMaps"). Multiple databases can be selected}

\item{annot_custom_db}{An optional dataframe from which the isomers will be sampled. It should have 2 columns : formula and molecule name. If provided, \code{annot_db} will be ignored.}

\item{endogenous_only}{A logical indicating whether to only consider endogenous metabolites (default = TRUE).}

\item{pathway_assoc_only}{A logical indicating whether to only consider metabolites associated with a biological pathway (default = FALSE)}

\item{remove_expected_predicted}{A logical indicating whether to remove expected and predicted isomers based on HMDB status (default = TRUE)}

\item{annotations}{An optional custom list of length n, with each element contains a vector of isomer names. If not specified,
S2IsoMEr uses the CoreMetabolome, LIPIDMAPS, SwissLipids, and HMDB databases from METASPACE (https://metaspace2020.eu/) to generate an annotation list automatically.}

\item{annotation.weights}{An optional list of length n, each element contains a vector of isomer weights. Only when annotations is provided as list.}

\item{consider_isomers}{A logical indicating whether to include isomers (default = TRUE)}

\item{consider_isobars}{A logical indicating whether to include isobars (default = FALSE)}

\item{mass_range_ppm}{A numeric indicating the mass range in ppm (default: mass_range_ppm = 3). Molecular formulas + adducts within this range will be treated as isobars. Only required when isobars = TRUE.}

\item{polarization_mode}{A character with either 'positive' (default) or 'negative'. Only required when isobars = TRUE. When set to 'positive', included adducts are '+H', '+Na', and '+K'.
When set to 'negative', included adducts are '-H', '+Cl'.}

\item{include}{An optional logical vector of length n indicating whether to include the annotations in the analysis.}

\item{molecule_type}{A character specifying the feature type for the enrichment background, either Metabolites or Lipids. Valid choices are c("Metabo", "Lipid").}

\item{background_type}{A character specifying the background type for enrichment, choose one from c("LION","main_class","super_class","sub_class","pathways").}

\item{custom_bg}{A named list with character vectors of metabolite names. Default is NULL}

\item{termsOfInterest}{A character containing 'selection' (for default LION-term selection), 'all', or a vector of term names (see 'pathway').}

\item{condition.x}{first condition identifier for pairwise comparison.}

\item{condition.y}{second condition identifier for pairwise comparison.}

\item{ranking.by}{A character of either 'wilcox.test' (default), 't.test', 'logFC' or 'BWS', to rank metabolites using the respective statistic. Check \code{\link{rankScore}} for more details.
Ignored if \code{enrichment_type} is 'ORA'.}

\item{gsea.method}{A character of either 'ks_signed' or 'fgsea'. Ignored if \code{enrichment_type} is 'ORA'.}
}
\value{
An object of class S2IsoMEr.
}
\description{
initEnrichment() creates object to perform bootstrapping metabolite set enrichment analysis
}
\examples{
myTestRun <-
initEnrichment(scmatrix = scMatrix,
                   annotations = my_annotations,
                   annotation.weights = my_weights,
                   conditions = my_conditions,
                   condition.x = "A",
                   condition.y = "B" )


}
