% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Load_background}
\alias{Load_background}
\title{Load enrichment background}
\usage{
Load_background(
  mol_type = c("Lipid", "Metabo"),
  bg_type = c("LION", "main_class", "super_class", "sub_class", "pathways"),
  feature_type = c("sf", "name")
)
}
\arguments{
\item{mol_type}{Character indicating molecule type. Can be either \code{Lipid} for Lipids and \code{Metabo} for small molecules}

\item{bg_type}{Character indicating background type :
\itemize{
\item{\verb{LION : }}{LION ontology for Lipids only}
\item{\verb{main_class : }}{Level 1 classification - finer classification compared to super_class}
\item{\verb{super_class : }}{Level 0 classification}
\item{\verb{sub_class : }}{Level 2 classification - finer classification compared to main_class}
\item{\verb{pathways : }}{Biological pathways based upon KEGG, Reactome, and SMPDB}
}}

\item{feature_type}{Character indicating input feature type. Can be either \code{sf} for sum formula or ion
and \code{name} for molecule name.}
}
\value{
List where each element name correspond to a given term and values are
either metabolite names or sum formulas.
}
\description{
Load_background() retrieves a list of terms and associated metabolites for enrichment based on requested
molecule type, background type and input feature type.
}
\examples{
myTestRun <-
Load_background(mol_type = "Metabo",
                   bg_type = "main_class",
                   feature_type = "name")


}
