% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plot_MSEA_Multi_cond}
\alias{plot_MSEA_Multi_cond}
\title{Plot Heatmap of MSEA Results Across Multiple Conditions}
\usage{
plot_MSEA_Multi_cond(combined_MSEA_res, alpha_cutoff = 0.05)
}
\arguments{
\item{combined_MSEA_res}{A data frame containing combined MSEA results for each pairwise comparison from either enrichment results in \code{\link{Run_bootstrap_MSEA}} or \code{\link{Run_simple_MSEA}}}

\item{alpha_cutoff}{A numeric value specifying the cutoff for significance. Terms with p-values below this threshold will be marked as significant. Default is 0.05.}
}
\value{
A heatmap plot showing the NES values for each term across different condition pairs. Significant results are highlighted with an asterisk.
}
\description{
This function generates a heatmap of combined MSEA results. The heatmap displays normalized enrichment scores (NES) across multiple conditions and highlights significant results based on a specified alpha cutoff.
}
\details{
The function checks for the presence of required columns and ensures there are enough conditions to generate a meaningful plot. The heatmap includes color coding for NES values, with significant results indicated by asterisks.
}
\examples{
\dontrun{
data("example_MSEA_multicond")
p = plot_MSEA_Multi_cond(combined_MSEA_res = example_MSEA_multicond,
                         alpha_cutoff = 0.05)
}

}
