% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEA_utils.R
\name{rankScore}
\alias{rankScore}
\title{Rank metabolites for S2IsoMEr enrichment object}
\usage{
rankScore(object, ranking.by = "wilcox.test", alternative = "greater")
}
\arguments{
\item{object}{A S2IsoMEr object.}

\item{ranking.by}{A character string specifying the ranking method. Options include:
\itemize{
\item \code{"wilcox.test"} (default) - Ranks metabolites based on the results of a Wilcoxon test.
\item \code{"t.test"} - Ranks metabolites based on the results of a t-test.
\item \code{"logFC"} - Ranks metabolites based on log fold change.
\item \code{"BWS"} - Ranks metabolites based on BWS (Baumgartner-Weiss-Schindler ).
}}

\item{alternative}{A character string specifying the alternative hypothesis for statistical tests. Options are:
\itemize{
\item \code{"two.sided"} - Test for differences in both directions.
\item \code{"less"} - Test if the second condition is less than the first
\item \code{"greater"} - Test if the second condition is greater than the first
}
Default is \code{"greater"}.}
}
\value{
An object of class S2IsoMEr.
}
\description{
rankScore() ranks metabolites of S2IsoMEr object to perform bootstrapping metabolite set enrichment analysis
}
\examples{
\dontrun{
data("example_ORA_obj")
example_ORA_obj <-
rankScore(object = example_ORA_obj, ranking.by = 't.test')
}


}
