% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORA.R
\name{Run_bootstrap_ORA}
\alias{Run_bootstrap_ORA}
\title{Wrapper function for Bootstrap Over-Representation Analysis (ORA)}
\usage{
Run_bootstrap_ORA(
  marker_list,
  background,
  custom_universe = NULL,
  alpha_cutoff = 0.05,
  min_intersection = 3,
  consider_isobars = T,
  polarization_mode = NA,
  mass_range_ppm = 3,
  annot_db = "HMDB",
  annot_custom_db = NULL,
  use_LION = F,
  endogenous_only = T,
  pathway_assoc_only = F,
  remove_expected_predicted = T,
  annot_list = NULL,
  annot_weights = NULL,
  n_bootstraps = 50,
  boot_fract_cutoff = 0.5,
  q.val_cutoff = 0.2,
  selected_terms = NULL,
  adjust_contingency = T,
  report_ambiguity_scores = F
)
}
\arguments{
\item{marker_list}{A vector or list of marker sets to analyze.}

\item{background}{A list of named terms where each term is a character vector of metabolites.}

\item{custom_universe}{An optional vector specifying a custom universe of terms. If not provided all metabolites in background will be used as universe.}

\item{alpha_cutoff}{A numeric value indicating the alpha cutoff for significance.}

\item{min_intersection}{An integer specifying the minimum intersection required between input list and a given term in background.}

\item{consider_isobars}{A logical indicating whether to consider isobars.}

\item{polarization_mode}{A parameter for polarization mode. If provided, it will be used to consider isomers and isobars.}

\item{mass_range_ppm}{A numeric value for mass range in parts per million (ppm).}

\item{annot_db}{A character string specifying the annotation database(s) used for annotation. Check  \code{\link{build_iso_bg}} for more information.}

\item{annot_custom_db}{An optional custom annotation database. If provided, it will be used alongside or instead of the specified annotation database.}

\item{use_LION}{A logical indicating whether to use LION ontology for select isomers/isobars for background.}

\item{endogenous_only}{A logical indicating whether to consider only endogenous compounds. Applies only for HMDB and CoreMetabolome annotation databases.}

\item{pathway_assoc_only}{A logical indicating whether to consider only pathway-associated compounds. Applies only for HMDB and CoreMetabolome annotation databases.}

\item{remove_expected_predicted}{A logical indicating whether to remove expected and predicted annotations. Applies only for HMDB and CoreMetabolome annotation databases.}

\item{annot_list}{An optional list of annotations. If provided, it will be used instead of generating isomers from built-in formula to molecule mapping.}

\item{annot_weights}{An optional list of annotation weights. If provided, it will be used in the ambiguity score calculation and bootstrap analysis.}

\item{n_bootstraps}{An integer specifying the number of bootstrap iterations.}

\item{boot_fract_cutoff}{A numeric value specifying the minimum bootstrap fraction cutoff required. Default to 0.5 .}

\item{q.val_cutoff}{A numeric value specifying the q-value cutoff for significance.}

\item{selected_terms}{An optional vector of selected terms to focus on. If provided, the analysis will be restricted to these terms.}

\item{adjust_contingency}{A logical indicating whether to adjust the contingency table to account for isomeric ambiguity.}

\item{report_ambiguity_scores}{A logical indicating whether to report ambiguity scores. If TRUE, ambiguity scores will be included in the results.}
}
\value{
A list containing the ORA results for each group of markers. Both filtered and per-bootstrap results are provided.
}
\description{
This function performs a bootstrap-based over-representation analysis (ORA) on a given list of metabolites and background.
}
\examples{
\dontrun{
marker_list <- c("B1", "B3", "B5")
background <- list(
  "Term1" = c("B1", "B2", "B3"),
  "Term2" = c("B4", "B5", "B6")
)
results <- Run_bootstrap_ORA(marker_list, background)
}
}
