% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Run_enrichment}
\alias{Run_enrichment}
\title{Run Metabolite/Lipids enrichment analysis for single cell metabolomics}
\usage{
Run_enrichment(
  object,
  Run_DE = FALSE,
  DE_pval_cutoff = 0.05,
  DE_LFC_cutoff = 1,
  min.pct.diff = 0.1,
  ...
)
}
\arguments{
\item{object}{A numeric matrix of n metabolites (rows) and m cells or measurments (columns).}

\item{Run_DE}{A logical indicating whether to run differential analysis using limma's rank Sum Test With Correlation. Ignored if enrichment type is 'MSEA'.}

\item{DE_pval_cutoff}{A numeric indicating p-value cutoff. Default is 0.05.}

\item{DE_LFC_cutoff}{A numeric indicating the minimum log2 fold change for differential analysis}

\item{min.pct.diff}{A numeric indicating the minimum percentage difference between samples/cells in both conditions for a marker to be considered differentially abundant.}

\item{...}{Additional arguments passed to functions used internally, such as \code{\link{Run_bootstrap_ORA}},
\code{\link{Run_simple_ORA}}, \code{\link{Run_bootstrap_MSEA}}, and \code{\link{Run_simple_MSEA}}.
Consult the specific function documentation for details on these arguments.}
}
\value{
Data.frame with enrichment results
}
\description{
Run_enrichment() A wrapper to run either ORA or MSEA based on the initialized enrichment object
}
\examples{
myTestRun <-
Run_enrichment(object = object,
                   Run_DE = TRUE)


}
