% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2IsoMEr-package.R
\docType{package}
\name{S2IsoMEr-package}
\alias{S2IsoMEr}
\alias{S2IsoMEr-package}
\title{S2IsoMEr: Spatial and Single-Cell Isomeric Metabolite Enrichment in R}
\description{
R package performing Overrepresentation analysis (ORA) and Metabolite Set Enrichment Analysis (MSEA) on Spatial and single-cell metabolomics datasets using bootstrapping of isomeric/isobaric molecules.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://alexandrovteam.github.io/S2IsoMEr/}
}

}
\author{
\strong{Maintainer}: Bishoy Wadie \email{bishoy.wadie@embl.de} (\href{https://orcid.org/0000-0001-8692-1107}{ORCID})

Authors:
\itemize{
  \item Martijn Molenaar \email{m.r.molenaar@protonmail.com} (\href{https://orcid.org/0000-0001-5221-608X}{ORCID})
}

Other contributors:
\itemize{
  \item Theodore Alexandrov \email{talexandrov@health.ucsd.edu} (\href{https://orcid.org/0000-0001-9464-6125}{ORCID}) [copyright holder, funder]
}

}
\keyword{internal}
