% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setConditions}
\alias{setConditions}
\title{Set conditions for enrichment analysis}
\usage{
setConditions(object, condition.x = NULL, condition.y = NULL)
}
\arguments{
\item{object}{A S2IsoMEr object.}

\item{condition.x}{A optional character describing the reference condition.}

\item{condition.y}{A optional character describing condition to interest.}
}
\value{
An object of class \code{S2IsoMEr}.
}
\description{
Set conditions for enrichment analysis
}
\examples{
\dontrun{
data("example_MSEA_obj")
example_MSEA_obj = setConditions(example_MSEA_obj, condition.x = 'U', condition.y = "FI")
}

}
