% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_fgsea.R
\name{simple_fgsea}
\alias{simple_fgsea}
\title{Run simple fGSEA enrichment}
\usage{
simple_fgsea(
  pathways,
  stats,
  minSize = 1,
  maxSize = length(stats) - 1,
  eps = 1e-50,
  scoreType = c("std", "pos", "neg"),
  nPermSimple = 1000
)
}
\arguments{
\item{pathways}{List of metabolite sets to check.}

\item{stats}{Named vector of metabolite-level stats. Names should be the same as in 'pathways'}

\item{minSize}{Minimal size of a metabolite set to test. All pathways below the threshold are excluded.}

\item{maxSize}{Maximal size of a metabolite set to test. All pathways above the threshold are excluded.}

\item{eps}{This parameter sets the boundary for calculating the p value.}

\item{scoreType}{This parameter defines the GSEA score type. Possible options are ("std", "pos", "neg")}

\item{nPermSimple}{Number of permutations in the simple fgsea implementation for preliminary estimation of P-values}
}
\value{
A table with GSEA results. Each row corresponds to a tested pathway.
The columns are the following:
\itemize{
\item pathway -- name of the pathway as in \code{names(pathway)};
\item pval -- an enrichment p-value;
\item padj -- a BH-adjusted p-value;
\item ES -- enrichment score, same as in Broad GSEA implementation;
\item NES -- enrichment score normalized to mean enrichment of random samples of the same size;
\item nMoreExtreme\verb{-- a number of times a random gene set had a more extreme enrichment score value; \\item size -- size of the pathway after removing genes not present in}names(stats)`.
\item leadingEdge -- vector with indexes of leading edge genes that drive the enrichment, see \url{http://software.broadinstitute.org/gsea/doc/GSEAUserGuideTEXT.htm#_Running_a_Leading}.
}
}
\description{
simple_fgsea calls a simple fGSEA from the fgsea package
}
\examples{
bg = list("Term_A" = paste0("mol_", sample(100,20)),"Term_B" = paste0("mol_", sample(100,20)))
mol_ranks = runif(10, -3, 3)
names(mol_ranks) = sample(unique(unlist(bg)), 10)
mol_ranks = mol_ranks[order(mol_ranks)]
simple_fgsea(pathways = bg, stats = mol_ranks, minSize = 3, scoreType = "std")
}
